+++ src/app/code.component.ts
@@ -1,11 +1,5 @@
 import { Component, ElementRef, AfterViewInit, Input, ViewChild } from '@angular/core';
 
-declare global {
-  interface Window {
-    Prism: any;
-  }
-}
-
 @Component({
   selector: 'app-code',
   template: `
@@ -20,9 +14,9 @@ export class CodeComponent implements AfterViewInit {
 
   constructor(public el: ElementRef) {}
 
-  ngAfterViewInit(): void {
-    if (window.Prism) {
-      window.Prism.highlightElement(this.codeViewChild.nativeElement);
+  ngAfterViewInit() {
+    if (window['Prism']) {
+      window['Prism'].highlightElement(this.codeViewChild.nativeElement);
     }
   }
 }