+++ src/app/service/product.service.ts
@@ -8,11 +8,13 @@ import { Product } from '../domain/model';
 export class ProductService {
   constructor(private http: HttpClient) {}
 
-  getProductsSmall(): Promise<Product[]> {
+  getProductsSmall() {
     return this.http
       .get<any>('assets/demo/data/products-small.json')
       .toPromise()
-      .then((res) => res.data as Product[])
-      .then((data) => data);
+      .then((res) => <Product[]>res.data)
+      .then((data) => {
+        return data;
+      });
   }
 }